% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.squared.R
\name{cosine.squared}
\alias{cosine.squared}
\title{Cosine squared}
\usage{
cosine.squared(b1, b2)
}
\arguments{
\item{b1}{First vector}

\item{b2}{Second vector}
}
\value{
The cosine squared of the two vecotrs
}
\description{
Calculates the quality of the correlation between two vectors (real beta and estimated beta)
}
\examples{
b1 = c(-0.22,-0.31,-0.27)
b2 = c(1,1,1)
cosine.squared(b1,b2)
}
