% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_squared.R
\name{cosine_squared}
\alias{cosine_squared}
\title{Cosine squared}
\usage{
cosine_squared(b1, b2)
}
\arguments{
\item{b1}{First vector}

\item{b2}{Second vector}
}
\value{
The cosine squared between the two vectors
}
\description{
Calculates the quality of the correlation between two vectors
(real beta and estimated beta)
}
