% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{do_hard_thresholding}
\alias{do_hard_thresholding}
\title{Hard thresholding}
\usage{
do_hard_thresholding(M, lambda)
}
\arguments{
\item{M}{The matrix that will be thresholded}

\item{lambda}{The parameter of the tresholding}
}
\value{
The matrix M thresholded
}
\description{
Apply a hard thresholding of parameter lambda to the matrix M
}
