% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_matrix}
\alias{ensure_matrix}
\title{Ensure matrix}
\usage{
ensure_matrix(data)
}
\arguments{
\item{data}{The data}
}
\value{
The data as a matrix
}
\description{
Verify if `data` is a matrix. If not, it is converted into
a matrix if possible. Else, an error is thrown.
}
