% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SIR_threshold_opt.R
\name{plot.SIR_threshold_opt}
\alias{plot.SIR_threshold_opt}
\title{Graphical output of SIR_threshold_opt}
\usage{
\method{plot}{SIR_threshold_opt}(x, choice = "", ...)
}
\arguments{
\item{x}{A SIR_threshold_opt object}

\item{choice}{the graph to plot: 
\itemize{
  \item "estim_ind" Plot the estimated index by the SIR model versus Y
  \item "opt_lambda" Plot the choice of the optimal lambda
  \item "cos2_selec" Plot the evolution of cos^2 and variable selection according to 
  lambda
  \item "regul_path" Plot the regularization path of b
  \item "" Plot every graphs
}}
}
\description{
Display the 10 first eigen values,the estimated index versus Y of the SIR model,
the evolution of cos^2 and variable selection according to lambda, and the 
regularization path of b
}
\examples{
# Generate Data
set.seed(10)
n <- 200
beta <- c(1,1,rep(0,8))
X <- mvtnorm::rmvnorm(n,sigma=diag(1,10))
eps <- rnorm(n)
Y <- (X\%*\%beta)**3+eps

# Apply SIR with hard thresholding
SIR_threshold_opt(Y,X,H=10,n_lambda=300,thresholding="hard")

# Apply SIR with soft thresholding
SIR_threshold_opt(Y,X,H=10,n_lambda=300,thresholding="soft")

# Estimated index versus Y
plot.SIR_threshold_opt(res,choice="estim_ind")

# Choice of optimal lambda
plot.SIR_threshold_opt(res,choice="opt_lambda")

# Evolution of cos^2 and var selection according to lambda
plot.SIR_threshold_opt(res,choice="cos2_selec")

# Regularization path
plot.SIR_threshold_opt(res,choice="regul_path")
}
