% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{SIR}
\alias{SIR}
\title{Vanilla SIR}
\usage{
SIR(Y, X, H = 10, graphic = TRUE)
}
\arguments{
\item{Y}{A numeric vector representing the dependent variable (a response vector).}

\item{X}{A matrix representing the quantitative explanatory variables (bind by column).}

\item{H}{The chosen number of slices.}
}
\value{
An object of class SIR, with attributes:
\item{beta}{This is an estimated EDR direction, which is the principal 
eigenvector of the interest matrix.}
\item{M1}{The interest matrix.}
\item{eig.val}{The eigenvalues of the interest matrix.}
\item{eig.vect}{A matrix corresponding to the eigenvectors of the interest matrix.}
\item{n}{Sample size.}
\item{p}{The number of variables in X.}
}
\description{
Apply a single-index SIR (Sliced Inverse Regression) on (X,Y) with H slices.
}
\examples{
# Generate Data
set.seed(10)
n <- 500
beta <- c(1,1,rep(0,8))
X <- mvtnorm::rmvnorm(n,sigma=diag(1,10))
eps <- rnorm(n)
Y <- (X\%*\%beta)**3+eps

# Apply SIR
SIR(Y,X,H=10)
}
