% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.threshold.opt.R
\name{SIR.threshold.opt}
\alias{SIR.threshold.opt}
\title{SIR optimally thresholded}
\usage{
SIR.threshold.opt(
  Y,
  X,
  H = 10,
  N.lambda = 100,
  thresholding = "hard",
  graphic = TRUE,
  output = TRUE
)
}
\arguments{
\item{Y}{A numeric vector representing the dependent variable (a response vector).}

\item{X}{A matrix representing the quantitative explanatory variables (bind by column).}

\item{H}{The chosen number of slices.}

\item{N.lambda}{The number of lambda to test. The N.lambda tested lambdas are 
between 0 and the maximum value of the interest matrix.}

\item{thresholding}{The thresholding method (choose between hard, soft)}

\item{graphic}{A boolean, set to TRUE to plot graphs}

\item{output}{A boolean, set to TRUE to print informations}
}
\value{
An object of class SIR.threshold.opt, with attributes:
\item{b.opt}{This is the optimal estimated EDR direction, which is the principal 
eigenvector of the interest matrix.}
\item{lambdas}{A vector that contains the tested lambdas}
\item{lambda.opt}{The optimal lambda}
\item{mat.b.th}{A matrix of size p*N.lambda that contains an estimation of beta 
in the columns for each lambda}
\item{N.lambda}{The number of lambda tested}
\item{vect.nb.zeros}{The number of 0 in the estimation of the vector beta}
\item{list.relevant.variables}{A list that contains the variables selected by 
the model}
}
\description{
Apply a single-index SIR (Sliced Inverse Regression) on (X,Y) with H slices, 
with thresholding of the matrix of interest by the *optimal* lambda parameter.
}
\examples{

# Generate Data
set.seed(10)
n <- 200
beta <- c(1,1,rep(0,8))
X <- mvtnorm::rmvnorm(n,sigma=diag(1,10))
eps <- rnorm(n)
Y <- (X\%*\%beta)**3+eps

# Apply SIR with hard thresholding
SIR.threshold.opt(Y,X,H=10,N.lambda=300,thresholding="hard",graphic=TRUE,output=TRUE)

# Apply SIR with soft thresholding
SIR.threshold.opt(Y,X,H=10,N.lambda=300,thresholding="soft",graphic=TRUE,output=TRUE)
}
