% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_bootstrap.R
\name{SIR_bootstrap}
\alias{SIR_bootstrap}
\title{Bootstrap SIR}
\usage{
SIR_bootstrap(Y, X, H = 10, B = 10)
}
\arguments{
\item{Y}{A numeric vector representing the dependent variable (a response vector).}

\item{X}{A matrix representing the quantitative explanatory variables (bind by column).}

\item{H}{The chosen number of slices.}

\item{B}{The number of bootstrapped sample to draw}
}
\value{
An object of class SIR.bootstrap, with attributes:
\item{beta}{This is an estimated EDR direction, which is the principal eigenvector
of the interest matrix.}
\item{mat.b.est}{A matrix of size p*B that contains an estimation of beta in 
the columns for each bootstrapped sample}
\item{n}{Sample size.}
\item{p}{The number of variables in X.}
}
\description{
Apply a single-index SIR (Sliced Inverse Regression) on B bootstraped sample of
}
\examples{
# Generate Data
set.seed(10)
n <- 500
beta <- c(1,1,rep(0,8))
X <- mvtnorm::rmvnorm(n,sigma=diag(1,10))
eps <- rnorm(n)
Y <- (X\%*\%beta)**3+eps

# Apply bootstrap SIR
SIR_bootstrap(Y,X,H=10,B=10)
}
