% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_threshold.R
\name{SIR_threshold}
\alias{SIR_threshold}
\title{SIR thresholded}
\usage{
SIR_threshold(Y, X, H = 10, lambda = 0, thresholding = "hard")
}
\arguments{
\item{Y}{A numeric vector representing the dependent variable (a response vector).}

\item{X}{A matrix representing the quantitative explanatory variables (bind by column).}

\item{H}{The chosen number of slices.}

\item{lambda}{The thresholding parameter}

\item{thresholding}{The thresholding method (choose between hard, soft)}
}
\value{
An object of class SIR.threshold, with attributes:
\item{beta}{This is an estimated EDR direction, which is the principal eigenvector 
of the interest matrix.}
\item{M1_th}{The interest matrix thresholded.}
\item{eig.val}{The eigenvalues of the interest matrix thresholded.}
\item{eig.vect}{A matrix corresponding to the eigenvectors of the interest matrix.}
\item{n}{Sample size.}
\item{p}{The number of variables in X.}
\item{nb.zeros}{The number of 0 in the estimation of the vector beta}
\item{list.relevant.variables}{A list that contains the variables selected by the
model}
\item{cos.squared}{The cosine squared between vanilla SIR and SIR thresholded}
}
\description{
Apply a single-index SIR (Sliced Inverse Regression) on (X,Y) with H slices, with 
thresholding of the matrix of interest by the lambda parameter.
}
\examples{

# Generate Data
set.seed(10)
n <- 500
beta <- c(1,1,rep(0,8))
X <- mvtnorm::rmvnorm(n,sigma=diag(1,10))
eps <- rnorm(n)
Y <- (X\%*\%beta)**3+eps

# Apply SIR with hard thresholding
SIR_threshold(Y,X,H=10,lambda=0.2,thresholding="hard")

# Apply SIR with soft thresholding
SIR_threshold(Y,X,H=10,lambda=0.2,thresholding="soft")
}
